# image/png: optimise RGBA encoding