--- a/src/image/png/writer.go
+++ b/src/image/png/writer.go
@@ -450,22 +450,27 @@ func (e *encoder) writeImage(w io.Writer, m image.Image, cb int, level int) erro
 			if nrgba != nil {
 				offset := (y - b.Min.Y) * nrgba.Stride
 				copy(cr[0][1:], nrgba.Pix[offset:offset+b.Dx()*4])
+			} else if rgba != nil {
+				for x := b.Min.X; x < b.Max.X; x++ {
+					offset := rgba.PixOffset(x, y)
+					a := rgba.Pix[offset+3]
+					if a == 0 {
+						// Do nothing, keep next 4 bites as 0, it's a transparent pixel
+					} else if a == 0xff {
+						// opaque pixel, no need to un-premultiply
+						copy(cr[0][i:i+4], rgba.Pix[rgba.PixOffset(x, y):rgba.PixOffset(x, y)+4])
+					} else {
+						c := color.NRGBAModel.Convert(m.At(x, y)).(color.NRGBA)
+						cr[0][i+0] = c.R
+						cr[0][i+1] = c.G
+						cr[0][i+2] = c.B
+						cr[0][i+3] = c.A
+					}
+					i += 4
+				}
 			} else {
 				// Convert from image.Image (which is alpha-premultiplied) to PNG's non-alpha-premultiplied.
 				for x := b.Min.X; x < b.Max.X; x++ {
-					if rgba != nil {
-						// If it's RGBA, fully opaque and fully transparent pixels are two common cases to be optimized
-						a := rgba.Pix[rgba.PixOffset(x, y)+3]
-						if a == 0 { // keep next 4 bites as 0, it's a transparent pixel
-							i += 4
-							continue
-						}
-						if a == 0xff { // opaque pixel, no need to un-premultiply
-							copy(cr[0][i:i+4], rgba.Pix[rgba.PixOffset(x, y):rgba.PixOffset(x, y)+4])
-							i += 4
-							continue
-						}
-					}
 					c := color.NRGBAModel.Convert(m.At(x, y)).(color.NRGBA)
 					cr[0][i+0] = c.R
 					cr[0][i+1] = c.G