@alx-ef 2022-09-17
## Cove coverage
### Before
```
ok  	std/image/png	0.702s	coverage: 85.3% of statements
```
### After 
```
ok  	std/image/png	0.791s	coverage: 86.2% of statements
```

## Benchmark results
### Before
```
/usr/local/go/bin/go test -v ./... -bench . -run ^$
goos: darwin
goarch: arm64
pkg: std/image/png
 . . .
BenchmarkEncodeRGBA-10           100	  10812776 ns/op	 113.64 MB/s	 3321754 B/op	  614432 allocs/op
PASS
ok  	std/image/png	19.135s
```

### After optimizations
Almost 3x faster
```
BenchmarkEncodeRGBA-10           301	   3980006 ns/op	 308.74 MB/s	  864128 B/op	      32 allocs/op
```
### After benchmark reworking
Actually, the original benchmark tested only one corner case, so I reworked BenchmarkEncodeRGBA, making it closer to real-world usage. Thus, the benchmark became much slower 
```
BenchmarkEncodeRGBA-10            46	  25082351 ns/op	  48.99 MB/s	 1084532 B/op	   55117 allocs/op
```
but anyway, it's still faster than without optimizations:
```
BenchmarkEncodeRGBA-10    	      34	  30837667 ns/op	  39.85 MB/s	 3321862 B/op	  614436 allocs/op
```